Provide start_slam + stop_slam for rc_dynamics
Adapt to newest rc_dynamics version.
Changed interface:
  start(bool with_restart) -> start() and restart()

This way it is consistent with the REST-API.

With a straightforward extension, some combinations of with_restart and
starting SLAM would not be directly translatable to (and thus not be
consistent with) the REST API inteface, e.g. with
- start(bool with_restart, bool with_slam)
start(true, true) needs REST-API "restart" if SLAM is running, but
"restart" and "start_slam" otherwise.